Rectangle abstracted as an interface
